overhaul of call-test for some hacky multihost infra integration
