+++ .github/workflows/call-test-multihost-dualt3k.yml
@@ -199,13 +199,13 @@ jobs:
           export TT_METAL_LOGGER_TYPES=Fabric,Always
           # export TT_RUNTIME_USING_DUALT3K=1
 
-          echo "--- topology env export (before eval) ---"
-          python -m tests.torch.multi_host.topology \
+          echo "--- setting up multihost topology env ---"
+          TOPO_EXPORTS=$(python scripts/multihost_topology.py \
             --topology dual_t3k \
-            --script-dir tests/torch/multi_host/experimental
-          eval $(python -m tests.torch.multi_host.topology \
-            --topology dual_t3k \
-            --script-dir tests/torch/multi_host/experimental)
+            --script-dir tests/torch/multi_host/experimental) \
+            || { echo "topology setup failed: $TOPO_EXPORTS" >&2; exit 1; }
+          echo "$TOPO_EXPORTS"
+          eval "$TOPO_EXPORTS"
 
           pytest -svv tests/runner/test_models.py::test_all_models_torch[gpt_oss/pytorch-20B-tensor_parallel-inference]
         '
